riscv128 : updating configuration files
Setting up the configuration files in such a way that the support for
building a riscv128 cross-environment tool can be bootstrapped.
Once the patch applied the configuration can be launched smoothly:
../configure --target=riscv128-unknown-elf --enable-maintainer-mode
Updated with upstream.

configure: regenerate rv128 compilation files

bfd: Addition of 128-bit support in elf file format and bfd

Supporting 128-bit architectures means having a dedicated elf128 file
format and then a bunch of somehow little modifications spread around
many files.
The elf128 data structures are very similar to the 64-bit ones, and use
the same trick to ensure alignement of the 128-bit integers: moving
fields around to ensure packed structs with as few dummy fields as
possible.
Still a few FIXME around to check, as certain values might well depend
on the data-size.

bfd/cpu-riscv.c: make rv128 the default for now

binutils/readelf.c: casts for proper compilation

Because the binutils guys removed the bfd_vma type and print function
during the summer, hackish solution to output 64 lower bits of vma on
warning and error messages.

fix: print size in readelf

rv128: Adding support in gas

The new instructions have to be known from gas, along with their
specificities. Part of this is close from the modifications done
in opcodes.

riscv128 : Add ld support

Making what's necessary to have a 128-bit riscv linker.
In essence:
  * changing Makefile.am
  * adding emulation scripts

+++ bfd/elf-attrs.c
@@ -2894,7 +2894,7 @@ oav2_parse_section (bfd *abfd,
 	{
 	  _bfd_error_handler
 	    (_("%pB: error: could not parse subsection at offset %" PRIx64),
-	     abfd, hdr->sh_size - remaining);
+	     abfd, (uint64_t) hdr->sh_size - remaining);
 	  bfd_set_error (bfd_error_wrong_format);
 	  break;
 	}
